wrap i18n in a class
