+++ keyboards/index.php
@@ -19,7 +19,7 @@
     'language' => isset($_SESSION['lang']) ? $_SESSION['lang'] : 'en',
     'css' => [Util::cdn('css/template.css'), Util::cdn('keyboard-search/search.css')],
     'js' => [Util::cdn('keyboard-search/jquery.mark.js'), Util::cdn('keyboard-search/dedicated-landing-pages.js'),
-      Util::cdn('js/i18n/i18n.js'),
+      Util::cdn('js/i18n/i18n.mjs'),
       Util::cdn('keyboard-search/search.js')]
   ];
